Adds monitors to driver skeleton
- Adds logic to start monitors in Faban Driver skeleton according to
their run phase;
- Updates import regarding to test-definition-model, to account for its
refactoring;
- Updates Dockerfile to download collectors and monitors deployment
descriptors, plus monitors library;
