+++ Dockerfile.ci
@@ -15,6 +15,9 @@ ENV GENERATION_RESOURCES_ROOT /app/drivers
 ENV LIBRARIES_ROOT ${GENERATION_RESOURCES_ROOT}/libraries
 ENV PLUGINS_ROOT ${GENERATION_RESOURCES_ROOT}/plugins
 ENV TEMPLATES_ROOT ${GENERATION_RESOURCES_ROOT}/templates
+ENV BENCHFLOW_SERVICES_ROOT ${GENERATION_RESOURCES_ROOT}/benchflow-services
+ENV BENCHFLOW_COLLECTORS_ROOT ${BENCHFLOW_SERVICES_ROOT}/collectors
+ENV BENCHFLOW_MONITORS_ROOT ${BENCHFLOW_SERVICES_ROOT}/monitors
 
 COPY configuration.yml /app/
 COPY ./target/benchflow-drivers-maker.jar /app/
