+++ application/src/main/scala/cloud/benchflow/experiment/sources/processors/drivers/operations/wfms/WfMSStartDriverOperationsProcessor.scala
@@ -2,7 +2,8 @@ package cloud.benchflow.experiment.sources.processors.drivers.operations.wfms
 
 import cloud.benchflow.experiment.sources.processors._
 import cloud.benchflow.driversmaker.utils.env.DriversMakerEnv
-import cloud.benchflow.test.config.{Operation, WfMSStartDriver}
+import cloud.benchflow.test.config.sut.wfms.WfMSStartDriver
+import cloud.benchflow.test.config.Operation
 import cloud.benchflow.test.config.experiment.BenchFlowExperiment
 import com.sun.faban.driver.{Timing, BenchmarkOperation}
 import org.apache.commons.io.FilenameUtils
