+++ application/src/test/resources/app/drivers/templates/harness/wfms/WfMSBenchmark.java
@@ -34,7 +34,6 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
     private ApacheHC3Transport http;
     private DriverConfig driverConfig;
     private Map<String, String> modelsStartID;
-    private Map<String, ServiceInfo> serviceInfoMap;
     //this is already in defaultfabanbenchmark2
 //    protected ParamRepository params;
 
@@ -99,143 +98,6 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
              */
             addProperty(properties, "model_" + modelNum + "_name", modelName);
             addProperty(properties, "model_" + modelNum + "_startID", processDefinitionId);
-        }
-
-    }
-
-    private static class ServiceInfo {
-
-        private String serviceName;
-        private List<CollectorInfo> collectors;
-
-        public ServiceInfo(String name) {
-            serviceName = name;
-        }
-
-        public void addCollector(CollectorInfo collector) {
-            collectors.add(collector);
-        }
-
-        public String getName() {
-            return serviceName;
-        }
-
-        public List<CollectorInfo> getCollectors() {
-            return collectors;
-        }
-
-    }
-
-    private static class CollectorInfo {
-
-        private String collectorName;
-        private String collectorId;
-        private List<MonitorInfo> monitors;
-        private String startAPI;
-        private String stopAPI;
-
-        public CollectorInfo(String name, String id) {
-            collectorName = name;
-            collectorId = id;
-        }
-
-        public String getName() {
-            return collectorName;
-        }
-
-        public String getId() {
-            return collectorId;
-        }
-
-        public List<MonitorInfo> getMonitors() {
-            return monitors;
-        }
-
-        public String getStartAPI() {
-            return startAPI;
-        }
-
-        public String getStopAPI() {
-            return stopAPI;
-        }
-
-        public void setStartAPI(String start) {
-            startAPI = start;
-        }
-
-        public void setStopAPI(String stop) {
-            stopAPI = stop;
-        }
-
-        public void addMonitor(MonitorInfo monitor) {
-            monitors.add(monitor);
-        }
-
-    }
-
-    private static class MonitorInfo {
-
-        private String monitorName;
-        private String monitorId;
-
-        private String startAPI;
-        private String monitorAPI;
-        private String stopAPI;
-
-        private String runPhase;
-        private Map<String, String> params;
-
-        public MonitorInfo(String name, String id) {
-            monitorName = name;
-            monitorId = id;
-        }
-
-        public String getName() {
-            return monitorName;
-        }
-
-        public String getId() {
-            return monitorId;
-        }
-
-        public String getStartAPI() {
-            return startAPI;
-        }
-
-        public String getStopAPI() {
-            return stopAPI;
-        }
-
-        public String getMonitorAPI() {
-            return monitorAPI;
-        }
-
-        public void setStartAPI(String start) {
-            startAPI = start;
-        }
-
-        public void setStopAPI(String stop) {
-            stopAPI = stop;
-        }
-
-        public void setMonitorAPI(String monitor) {
-            monitorAPI = monitor;
-        }
-
-        public String getRunPhase() {
-            return runPhase;
-        }
-
-        public void setRunPhase(String phase) {
-            runPhase = phase;
-        }
-
-        public Map<String, String> getParams() {
-            return params;
-        }
-
-        public void setParams(Map<String, String> p) {
-            params = p;
         }
 
     }
@@ -325,8 +187,8 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
                 File modelFile = new File(modelPath);
                 String processDefinitionId = null;
 
-                //TODO: add with spoon
-//              processDefinitionId = plugin.deploy(modelFile).get(modelName);
+                //add with spoon
+                //processDefinitionId = plugin.deploy(modelFile).get(modelName);
 
                 logger.info("PROCESS DEFINITION ID: " + processDefinitionId);
 
@@ -353,56 +215,12 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
 
 
     /**
-     * Retrieves the address for a benchflow service from the deployment manager
-     */
-    private String benchFlowServiceAddress(String benchFlowServiceId, String privatePort) throws Exception {
-        String deploymentManagerPortsApi = benchFlowComposeAddress + "/projects/" +
-                trialId + "/port/" + benchFlowServiceId + "/" + privatePort;
-
-        HttpMethod getServiceAddress = new GetMethod(deploymentManagerPortsApi);
-        http.getHttpClient().executeMethod(getServiceAddress);
-        String serviceAddress = new String(getServiceAddress.getResponseBody(), "UTF-8");
-        getServiceAddress.releaseConnection();
-        return serviceAddress;
-    }
-
-
-    /**
      * Starts collectors and related monitors
      */
     @Override
     @StartRun
     public void start() throws Exception {
         super.start();
-        //callCollectorsApi("start");
-        String privatePort = driverConfig.getXPathValue("benchFlowServices/privatePort");
-        for(ServiceInfo serviceInfo: serviceInfoMap.values()) {
-
-            List<CollectorInfo> collectorInfoList = serviceInfo.getCollectors();
-            for(CollectorInfo collectorInfo: collectorInfoList) {
-
-                List<MonitorInfo> monitorInfoList = collectorInfo.getMonitors();
-                for(MonitorInfo monitorInfo: monitorInfoList) {
-
-                    //monitor at start of load
-                    if(monitorInfo.getRunPhase().equals("start")) {
-
-                        String monitorEndpoint = benchFlowServiceAddress(monitorInfo.getId(), privatePort);
-                        //TODO: get monitor interface from factory and run .run() method
-                    }
-
-                    else if(monitorInfo.getRunPhase().equals("all")) {
-                        String monitorEndpoint = benchFlowServiceAddress(monitorInfo.getId(), privatePort);
-                        //TODO: get monitor interface from factory and run .start() method
-                    }
-
-                }
-
-                String collectorEndpoint = benchFlowServiceAddress(collectorInfo.getId(), privatePort);
-                //TODO: call start api for collector
-            }
-
-        }
     }
 
     /**
@@ -410,33 +228,6 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
      */
     @PostRun
     public void postRun() throws Exception {
-
-        String privatePort = driverConfig.getXPathValue("benchFlowServices/privatePort");
-        for(ServiceInfo serviceInfo: serviceInfoMap.values()) {
-
-            List<CollectorInfo> collectorInfoList = serviceInfo.getCollectors();
-            for(CollectorInfo collectorInfo: collectorInfoList) {
-
-                List<MonitorInfo> monitorInfoList = collectorInfo.getMonitors();
-                for(MonitorInfo monitorInfo: monitorInfoList) {
-
-                    //monitor at end of load
-                    if(monitorInfo.getRunPhase().equals("end")) {
-                        String monitorEndpoint = benchFlowServiceAddress(monitorInfo.getId(), privatePort);
-                        //TODO: get monitor interface from factory and run .run() method
-                    } else if(monitorInfo.getRunPhase().equals("all")) {
-                        String monitorEndpoint = benchFlowServiceAddress(monitorInfo.getId(), privatePort);
-                        //TODO: get monitor interface from factory and run .monitor() and .stop() methods
-                    }
-
-                }
-
-                String collectorEndpoint = benchFlowServiceAddress(collectorInfo.getId(), privatePort);
-                //TODO: call stop api for collector
-            }
-
-        }
-
         undeploy();
     }
 
@@ -479,6 +270,8 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
         driverConfig.addConfigurationNode("sutConfiguration", "sutEndpoint", sutEndpoint);
     }
 
+
+
     /**
      * Setup benchmarkDir, http transport, and services info map
      */
@@ -487,104 +280,6 @@ public class WfMSBenchmark extends DefaultFabanBenchmark2 {
         this.benchmarkDir = Paths.get(RunContext.getBenchmarkDir());
         this.http = new ApacheHC3Transport();
         this.driverConfig = new DriverConfig();
-        this.serviceInfoMap = new HashMap<String, ServiceInfo>();
-        parseBenchmarkConfiguration();
-    }
-
-    /**
-     * Parse run.xml to build a service info map
-     */
-    private void parseBenchmarkConfiguration() {
-
-        NodeList services = driverConfig.getNode("benchFlowServices/servicesConfiguration").getChildNodes();
-
-        //iterate over every service in the configuration
-        for(int serviceIndex = 0; serviceIndex < services.getLength(); serviceIndex++) {
-
-            Element service = (Element) services.item(serviceIndex);
-            String serviceName = service.getAttribute("name");
-            ServiceInfo serviceInfo = new ServiceInfo(serviceName);
-
-            NodeList collectors = service.getElementsByTagName("collector");
-
-            //iterate over all collectors for the given service
-            for(int collectorIndex = 0; collectorIndex < collectors.getLength(); collectorIndex++) {
-
-                Element collector = (Element) collectors.item(collectorIndex);
-                String collectorName = collector.getAttribute("name");
-                String collectorId = collector.getElementsByTagName("id").item(0).getTextContent();
-
-                CollectorInfo collectorInfo = new CollectorInfo(collectorName, collectorId);
-
-                Element apiNode = (Element) collector.getElementsByTagName("api").item(0);
-
-                //find out start and stop API values
-                NodeList collectorAPIs = apiNode.getChildNodes();
-                for(int apiIndex = 0; apiIndex < collectorAPIs.getLength(); apiIndex++) {
-
-                    Element currentAPI = (Element)collectorAPIs.item(apiIndex);
-                    if(currentAPI.getTagName().equals("start")) {
-                        collectorInfo.setStartAPI(currentAPI.getTextContent());
-                    } else if(currentAPI.getTagName().equals("stop")) {
-                        collectorInfo.setStopAPI(currentAPI.getTextContent());
-                    }
-                }
-
-                NodeList monitors = collector.getElementsByTagName("monitors").item(0).getChildNodes();
-
-                for(int monitorIndex = 0; monitorIndex < monitors.getLength(); monitorIndex++) {
-
-                    Element monitor = (Element) monitors.item(monitorIndex);
-                    String monitorName = monitor.getAttribute("name");
-                    String monitorId = monitor.getElementsByTagName("id").item(0).getTextContent();
-
-                    MonitorInfo monitorInfo = new MonitorInfo(monitorName, monitorId);
-
-                    //build monitor parameters map
-                    Element monitorConfiguration = (Element) monitor.getElementsByTagName("configuration").item(0);
-                    NodeList monitorConfigurationParams = monitorConfiguration.getElementsByTagName("param");
-                    Map<String, String> params = new HashMap<String, String>();
-                    for(int paramIndex = 0; paramIndex < monitorConfigurationParams.getLength(); paramIndex++) {
-
-                        Element param = (Element) monitorConfigurationParams.item(paramIndex);
-                        String paramName = param.getAttribute("name");
-                        String paramValue = param.getTextContent();
-                        params.put(paramName, paramValue);
-
-                    }
-
-                    monitorInfo.setParams(params);
-
-                    Element monitorApiNode = (Element) monitor.getElementsByTagName("api").item(0);
-
-                    //find out monitor APIs
-                    NodeList monitorAPIs = monitorApiNode.getChildNodes();
-                    for(int monitorApiIndex = 0; monitorApiIndex < monitorAPIs.getLength(); monitorApiIndex++) {
-
-                        Element currentMonitorAPI = (Element) monitorAPIs.item(monitorApiIndex);
-                        if(currentMonitorAPI.getTagName().equals("start")) {
-                            monitorInfo.setStartAPI(currentMonitorAPI.getTextContent());
-                        } else if(currentMonitorAPI.getTagName().equals("monitor")) {
-                            monitorInfo.setMonitorAPI(currentMonitorAPI.getTextContent());
-                        } else if(currentMonitorAPI.getTagName().equals("stop")) {
-                            monitorInfo.setStopAPI(currentMonitorAPI.getTextContent());
-                        }
-
-                    }
-
-                    //check phase
-                    String runPhase = monitor.getElementsByTagName("runPhase").item(0).getTextContent();
-                    monitorInfo.setRunPhase(runPhase);
-
-                    collectorInfo.addMonitor(monitorInfo);
-                }
-
-                serviceInfo.addCollector(collectorInfo);
-            }
-
-            serviceInfoMap.put(serviceName, serviceInfo);
-        }
-
     }
 
 }
